requirements.txt
discord.py==1.2.5
tortoise-orm==0.15.5
aiosqlite==0.11.0
pyyaml==5.2
python-dateutil==2.8.1